Refactored Log methods, Log table view

+++ migrations/m0x100.php
@@ -28,8 +28,8 @@ class m0x100 extends \JustCoded\WP\ImageOptimizer\core\Migration {
 		$charset_collate = $wpdb->get_charset_collate();
 		$table_name      = $wpdb->prefix . models\Media::TABLE_IMAGE_STATS;
 		$sql             = "CREATE TABLE $table_name (
-			id           mediumint(9) NOT NULL AUTO_INCREMENT,
-			attach_id    mediumint(9) NOT NULL,
+			id           BIGINT(20) NOT NULL AUTO_INCREMENT,
+			attach_id    BIGINT(20) NOT NULL,
 			image_size   VARCHAR(255) NOT NULL,
 			bytes_before VARCHAR(255) NOT NULL,
 			bytes_after  VARCHAR(255) NOT NULL,
